+++ Changelog.md
@@ -6,6 +6,14 @@ The file was started with Version `0.4`.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.8] unreleased
+
+This is yet another round of review – maybe also introducing a few test tools
+
+### Fixed
+
+* the approximate Hessians no longer store and overwrite the point they are built with.
+
 ## [0.6.7] September 9, 2026
 
 ### Added